make fmt private
