+++ include/mechanism_configuration/mechanism_configuration.hpp
@@ -5,7 +5,6 @@
 #pragma once
 
 #include <mechanism_configuration/errors.hpp>
-#include <mechanism_configuration/format_compat.hpp>
 #include <mechanism_configuration/mechanism.hpp>
 #include <mechanism_configuration/parse.hpp>
 #include <mechanism_configuration/types.hpp>
