+++ src/v1/reactions/photolysis.cpp
@@ -3,10 +3,10 @@
 // SPDX-License-Identifier: Apache-2.0
 
 #include <mechanism_configuration/errors.hpp>
-#include <mechanism_configuration/format_compat.hpp>
 #include <mechanism_configuration/types.hpp>
 
 #include <detail/check_schema.hpp>
+#include <detail/error_format.hpp>
 #include <detail/v1/reaction_parsers.hpp>
 #include <detail/v1/type_parsers.hpp>
 #include <detail/v1/type_schema.hpp>
