+++ src/validate.cpp
@@ -2,8 +2,9 @@
 //                         University of Illinois at Urbana-Champaign
 // SPDX-License-Identifier: Apache-2.0
 
-#include <mechanism_configuration/format_compat.hpp>
 #include <mechanism_configuration/validate.hpp>
+
+#include "detail/error_format.hpp"
 
 #include <optional>
 #include <string>
